# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for G4 Car Service Drivers Application.",
        "title": "G4 Drivers API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "samllachp@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/admin/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get global platform statistics (users, applications, referrals)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get Admin Stats",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admin.AdminStats"
                        }
                    },
                    "500": {
                        "description": "Stats Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/admin/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get paginated list of all users and their application status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "List Users (Admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admin.UserListResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "DB Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/user/dashboard": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get user dashboard including referral stats and referal list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Get My Dashboard",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dashboard.UserDashboard"
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "admin.AdminStats": {
            "type": "object",
            "properties": {
                "new_users_last_week": {
                    "type": "integer"
                },
                "total_comfort": {
                    "type": "integer"
                },
                "total_luxury": {
                    "type": "integer"
                },
                "total_referrals": {
                    "type": "integer"
                },
                "total_users": {
                    "type": "integer"
                }
            }
        },
...
        "admin.AdminUserListItem": {
            "type": "object",
            "properties": {
                "driver_category": {
                    "type": "string"
                },
                "driver_status": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "referred_count": {
                    "type": "integer"
                },
                "role": {
                    "type": "string"
                }
            }
        },
...
        "admin.PaginationMeta": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "items_per_page": {
                    "type": "integer"
                },
                "total_items": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "admin.UserListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/admin.AdminUserListItem"
                    }
                },
                "meta": {
                    "$ref": "#/definitions/admin.PaginationMeta"
                }
            }
        },
        "dashboard.ProfileInfo": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "referral_code": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "dashboard.ReferralItem": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "description": "Nuevo: Foto de perfil",
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "description": "Nuevo: Nombre del conductor (si existe)",
                    "type": "string"
                },
                "joined_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "dashboard.ReferralStats": {
            "type": "object",
            "properties": {
                "current_page": {
                    "description": "Nuevo",
                    "type": "integer"
                },
                "items_per_page": {
                    "description": "Nuevo",
                    "type": "integer"
                },
                "total_pages": {
                    "description": "Nuevo: Paginación",
                    "type": "integer"
                },
                "total_referred": {
                    "type": "integer"
                }
            }
        },
        "dashboard.UserDashboard": {
            "type": "object",
            "properties": {
                "profile": {
                    "$ref": "#/definitions/dashboard.ProfileInfo"
                },
                "referral_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dashboard.ReferralItem"
                    }
                },
                "referral_stats": {
                    "$ref": "#/definitions/dashboard.ReferralStats"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}